Separate source files from tests, extract snippets via Bluehawk
