+++ astro/src/code-snippets/webauthn/base64url-to-buffer.js
@@ -1,0 +1,9 @@
+// :snippet-start: base64url-to-buffer
+function base64URLToBuffer(base64URL) {
+  const base64 = base64URL.replace(/-/g, '+').replace(/_/g, '/');
+  const padLen = (4 - (base64.length % 4)) % 4;
+  return Uint8Array.from(atob(base64.padEnd(base64.length + padLen, '=')), c => c.charCodeAt(0));
+}
+// :snippet-end:
+
+module.exports = { base64URLToBuffer }; // :remove:
